Stop tracking typechain factories in git
Add entire typechain folder to .gitignore. Previously specific files
were being whitelisted, which is why the LilVRGDA factory did not
appear (it was not in the whitelist in .gitignore).

I'm think it's better not to track these in git because new diffs are
generated everytime the contracts are compiled with hardhat and because
the whitelisting via .gitignore is not obvious.

Developers should generate these client side by running `npx hardhat
compile` from nouns-contracts package.